requirements.txt
pyyaml
requests
flask
flask-cors